Improve preprocessing of model formulations.
