+++ optimus/model/__init__.py
@@ -1,5 +1,6 @@
 from . import acoustics
 from .acoustics import create_default_model
 from .acoustics import create_acoustic_model
+from . import nested
 from .nested import create_default_nested_model
 from .nested import create_nested_model
