+++ optimus/model/common.py
@@ -247,206 +247,3 @@ def _vector_to_gridfunction(vector, spaces):
     ]
 
     return gridfunctions
-
-
-def _check_validity_formulation(
-    formulation, formulation_parameters, preconditioner, preconditioner_parameters
-):
-    """
-    Check if the specified formulation and preconditioner is a valid choice.
-
-    Parameters
-    ----------
-    formulation : str
-        The type of boundary integral formulation.
-    formulation_parameters : dict
-        The parameters for the boundary integral formulation.
-    preconditioner : str
-        The type of operator preconditioner.
-    preconditioner_parameters : dict
-        The parameters for the operator preconditioner.
-
-    Returns
-    -------
-    formulation_name : str
-        The name of the boundary integral formulation.
-    preconditioner_name : str
-        The name of the operator preconditioner.
-    model_parameters : dict
-        The parameters for the preconditioned boundary integral formulation.
-    """
-
-    if not isinstance(formulation, str):
-        raise TypeError(
-            "The boundary integral formulation needs to be specified as a string."
-        )
-    else:
-        formulation_name = formulation.lower()
-
-    if preconditioner is None:
-        preconditioner_name = "none"
-    elif not isinstance(preconditioner, str):
-        raise TypeError("The preconditioner needs to be specified as a string.")
-    else:
-        preconditioner_name = preconditioner.lower()
-
-    if formulation_parameters is None:
-        formulation_parameters = {}
-    elif not isinstance(formulation_parameters, dict):
-        raise TypeError(
-            "The parameters of the boundary integral formulation need to be "
-            "specified as a dictionary."
-        )
-
-    if preconditioner_parameters is None:
-        preconditioner_parameters = {}
-    elif not isinstance(preconditioner_parameters, dict):
-        raise TypeError(
-            "The parameters of the preconditioner need to be specified as a dictionary."
-        )
-
-    if formulation_name not in ["pmchwt"]:
-        raise NotImplementedError("Unknown boundary integral formulation type.")
-
-    if preconditioner_name in ["none", "mass"]:
-        prec_params = {}
-    elif preconditioner_name == "osrc":
-        prec_params = _process_osrc_parameters(preconditioner_parameters)
-    else:
-        raise NotImplementedError("Unknown preconditioner type.")
-
-    model_parameters = {**formulation_parameters, **prec_params}
-
-    return formulation_name, preconditioner_name, model_parameters
-
-
-def _process_osrc_parameters(preconditioner_parameters):
-    """
-    Process the parameters for the OSRC preconditioner.
-
-    If the OSRC parameter is not specified in the input,
-    the global parameter is used.
-    The OSRC parameters are:
-        - npade: number of Padé expansions
-        - theta: angle of the branch cut of the Padé series
-        - damped_wavenumber: damped wavenumber
-        - wavenumber: wavenumber
-
-    Parameters
-    ----------
-    preconditioner_parameters : dict, None
-        The parameters of the preconditioner.
-
-    Returns
-    -------
-    osrc_parameters : dict
-        The parameters of the OSRC preconditioner.
-    """
-
-    from optimus import global_parameters
-
-    global_params_osrc = global_parameters.preconditioning.osrc
-
-    if preconditioner_parameters is None:
-        preconditioner_parameters = {}
-
-    osrc_parameters = {}
-
-    if "npade" in preconditioner_parameters:
-        npade = preconditioner_parameters["npade"]
-        if not (isinstance(npade, int) and npade > 0):
-            raise TypeError(
-                "The number of Padé expansions for the OSRC operator needs to be "
-                "a positive integer."
-            )
-        osrc_parameters["osrc_npade"] = npade
-    else:
-        osrc_parameters["osrc_npade"] = global_params_osrc.npade
-
-    if "theta" in preconditioner_parameters:
-        theta = preconditioner_parameters["theta"]
-        if not isinstance(theta, (int, float)):
-            raise TypeError(
-                "The angle of the branch cut of the Padé series for the "
-                "OSRC operator needs to be a float."
-            )
-        osrc_parameters["osrc_theta"] = theta
-    else:
-        osrc_parameters["osrc_theta"] = global_params_osrc.theta
-
-    if "damped_wavenumber" in preconditioner_parameters:
-        k_damped = preconditioner_parameters["damped_wavenumber"]
-        if not (isinstance(k_damped, (int, float, complex)) or k_damped is None):
-            raise TypeError(
-                "The damped wavenumber for the OSRC operators needs to be "
-                "a complex number."
-            )
-        osrc_parameters["osrc_damped_wavenumber"] = k_damped
-    else:
-        osrc_parameters["osrc_damped_wavenumber"] = global_params_osrc.damped_wavenumber
-
-    if "wavenumber" in preconditioner_parameters:
-        k_osrc = preconditioner_parameters["wavenumber"]
-        if not isinstance(k_osrc, (int, float, complex, str)):
-            raise TypeError(
-                "The wavenumber for the OSRC operators needs to be a complex number "
-                "or a string."
-            )
-        elif isinstance(k_osrc, str):
-            if k_osrc not in ["int", "ext"]:
-                raise TypeError(
-                    "The wavenumber for the OSRC operators needs to be a complex "
-                    "number, or one of the labels 'int' and 'ext'."
-                )
-        osrc_parameters["osrc_wavenumber"] = k_osrc
-    else:
-        osrc_parameters["osrc_wavenumber"] = global_params_osrc.wavenumber
-
-    return osrc_parameters
-
-
-def _process_calderon_parameters(preconditioner_parameters):
-    """
-    Process the parameters for the Calderón preconditioner.
-
-    Default values are taken from the global parameters.
-    The Calderón parameters are:
-        - domain: domain of the Calderón operator (exterior or interior)
-
-    Parameters
-    ----------
-    preconditioner_parameters : dict, None
-        The parameters of the preconditioner.
-
-    Returns
-    -------
-    calderon_parameters : dict
-        The parameters of the Calderón preconditioner.
-    """
-
-    from optimus import global_parameters
-
-    global_params_calderon = global_parameters.preconditioning.calderon
-
-    if preconditioner_parameters is None:
-        preconditioner_parameters = {}
-
-    calderon_parameters = {}
-
-    if "domain" in preconditioner_parameters:
-        domain = preconditioner_parameters["domain"]
-        if not isinstance(domain, str):
-            raise TypeError(
-                "The domain of the Calderón preconditioner needs to be a string."
-            )
-        elif domain not in ("exterior", "interior"):
-            raise ValueError(
-                "The domain of the Calderón preconditioner needs to be either "
-                "'exterior' or 'interior'."
-            )
-        else:
-            calderon_parameters["domain"] = domain
-    else:
-        calderon_parameters["domain"] = global_params_calderon.domain
-
-    return calderon_parameters
